# Hide Tile and TileExtended implementation details inside Map structure