… src/map_type.h
… src/misc_gui.cpp
… src/object_map.h
… src/pathfinder/npf/npf.cpp
… src/rail.cpp
… src/rail_map.h
… src/road_map.cpp
… src/road_map.h
… src/saveload/afterload.cpp
… src/saveload/map_sl.cpp
… src/saveload/oldloader_sl.cpp
… src/saveload/station_sl.cpp
… src/saveload/waypoint_sl.cpp
… src/script/api/script_story_page.cpp
… src/station_cmd.cpp
… src/station_map.h
… src/tile_map.h
… src/town_map.h
… src/tree_map.h
… src/tunnel_map.h
… src/tunnelbridge_map.h
… src/void_map.h
… src/water_cmd.cpp
… src/water_map.h