## Motivation / Problem

Initial trigger was CodeQL complaining about short global variable names `_m` and `_me`. Then realizing that the whole idea of splitting `_m` and `_me` is basically arbitrary for performance reasons, so why would the user need to be aware of that.
Similarly CodeQL complains about functions defined in code blocks, so when I saw the same globals for map size related purposes being used in a similar way, I though about just creating a `Map` structure with all the size related information as well as the tiles, that provides simple access to the different fields of tiles.

With this reordering/refactoring where the different fields of the tiles are going to be stored becomes quite trivial.


## Description

Move map size related accessors to Map, as well abstracting away the Tile and TileExtended structs into Map accessors for the different fields.
Also some minor improvements (first few commits) to the code, so other more approprate functions are used instead of directly accessing the map.

The use/migrate changes have been mostly done by use of bash scripts. In case of conflicts undoing the changes and running the appropriate one(s) of the scripts below could help with conflict solution.
```
for i in `grep '_m\[' -R .|sed s/:.*//|uniq|grep -v map_func.h`; do sed "s/_me\[\([^]]*\)\]\.m\([0-9]\)/Map::m\2(\1)/g" -i $i; sed "s/_m\[\([^]]*\)\]\.m\([0-9]\)/Map::m\2(\1)/g" -i $i; sed "s/_m\[\([^]]*\)\]\.type/Map::type(\1)/g" -i $i; sed "s/_m\[\([^]]*\)\]\.height/Map::height(\1)/g" -i $i; done

for i in `grep ScaleByMapSize -R .|sed s/:.*//`; do sed s/ScaleByMapSize/Map::ScaleBySize/g -i $i; done
for i in `grep MapLog -R .|sed s/:.*//`; do sed s/MapLog/Map::Log/g -i $i; done
for i in `grep MapMax -R .|sed s/:.*//`; do sed s/MapMax/Map::Max/g -i $i; done
for i in `grep MapSize -R .|sed s/:.*//`; do sed 's/ MapSize/ Map::Size/g' -i $i; done
```

## Limitations

Breaks a lot of other patches.
Still no memory safety for the tile arrays; was thinking about using std::vector but decided against it for this PR, but that could be added relatively easily.


## Checklist for review

Some things are not automated, and forgotten often. This list is a reminder for the reviewers.
* The bug fix is important enough to be backported? (label: 'backport requested')
* This PR touches english.txt or translations? Check the [guidelines](https://github.com/OpenTTD/OpenTTD/blob/master/docs/eints.md)
* This PR affects the save game format? (label 'savegame upgrade')
* This PR affects the GS/AI API? (label 'needs review: Script API')
    * ai_changelog.hpp, gs_changelog.hpp need updating.
    * The compatibility wrappers (compat_*.nut) need updating.
* This PR affects the NewGRF API? (label 'needs review: NewGRF')
    * newgrf_debug_data.h may need updating.
    * [PR must be added to API tracker](https://wiki.openttd.org/en/Development/NewGRF/Specification%20Status)