@LordAro 2023-01-21
Would be good to see some profiling to make sure all the extra function calls aren't adding any significant overhead (both debug and release)