@rubidium42 2023-01-21
> Would be good to see some profiling to make sure all the extra function calls aren't adding any significant overhead (both debug and release)

Some statistics with `-v null -g wentbourne.sav`, GCC10 on Debian with an i5-3337U; release without asserts average of 9 runs:
Original: 64.9 seconds 
This: 65.3 seconds (+0.7%)
This with merged `TileBase + TileExtended`: 66.5 (+2.4%)
This with merged `TileBase + TileExtended` and pointer in `Tile`: 65.7 (+1.2%)
The last one can probably be improved more when usage of `Tile` gets pushed further into the call stack, so it needs to do fewer lookups.

With debug and average of 3 runs:
Original: 489 seconds
This: 574 seconds (+17%)
This with `FORCEINLINE` for `Tile`: 485 seconds (-0.6%)

So, with optimisation there is barely a noticable difference, although the current split in two arrays is slightly faster. By resolving the pointer and reusing that you can claw back some of the performance loss. 
Note: release with asserts is somewhere around 50% slower!

With debug it's definitely another story. This current version is significantly slower, and "only" enabling inlining of small functions (`-finline-small-functions`) does not seem to achieve something useful. However, "reintroducing" `FORCEINLINE` (https://meghprkh.github.io/blog/posts/c++-force-inline/) for use in such special and trivial wrapper cases brings it back to the original speed.
Note that I added `FORCEINLINE` after finding out that this version is significantly slower than the original, however I doubt that `FORCEINLINE` will have an appreciable effect on release builds, so I have not created statistics with that change.

I can imagine `FORCEINLINE` also being used in other cases of trivial wrapper functions where we just get hit by a huge penalty in performance due to not inlining at all. Though we should set up some rules for its use, like only for trivial functions that only return some value and that are called really often.