@glx22 2023-01-21
Release builds probably already inline trivial wrappers.