@rubidium42 2023-01-28
> I can imagine `FORCEINLINE` also being used in other cases of trivial wrapper functions where we just get hit by a huge penalty in performance due to not inlining at all. Though we should set up some rules for its use, like only for trivial functions that only return some value and that are called really often.

This got added with #10406, so now there's also no statistically significant regression on the speed of debug builds anymore.